Generalize event handling
Finit is now able to receive generic events on the form:

[~]<name>[:<val>]

Tilde (~) indicate that the event has been negated (removed). The
optional value can be used to do matching on particular states of an
event.

The same format is used in the event section of a service to specify
the events that are required for a service to start. Events are
separated by a comma and are ANDed together.

E.g. given the following stanzas:

service <net/l3,net/gw>  /sbin/openvpn -f wan-normal.conf
service <net/l3,~net/gw> /sbin/openvpn -f lan-backup.conf

Both tunnels require net/l3 to start, then depending on whether net/gw
is available or not either the wan or lan configuration will be
started and the other one will be stopped.